aula_08a_0/requirements.txt
numpy~=2.2.6
matplotlib~=3.10.6
opencv-python~=4.12.0.88
PyOpenGL~=3.1.10
PyOpenGL-accelerate
pygame~=2.6.1